Следующая версия проекта

+++ delete_product.php
@@ -1,14 +1,17 @@
 <?php
   require 'connect.php';
 
+  if (!isset($moderator)) {
+    die('Неверный запрос');
+  }
+
   $data = $_POST;
 
   if (isset($data['id'])) {
     $product_id = $data['id'];
 
     $query_delete_properties = pg_query_params($conn, 'DELETE FROM properties WHERE product_id = $1', Array($product_id));
-    $query_delete_refund = pg_query_params($conn, 'DELETE FROM refunds WHERE product_id = $1',  Array($product_id));
-    $query_delete_products_in_orders = pg_query_params($conn, 'DELETE FROM products_in_baskets WHERE product_id = $1', Array($product_id));
+    $query_delete_products_in_orders = pg_query_params($conn, 'DELETE FROM client_products WHERE product_id = $1', Array($product_id));
     $query_delete_products_in_baskets = pg_query_params($conn, 'DELETE FROM products_in_orders WHERE product_id = $1', Array($product_id));
 
     $query_delete_product = pg_query_params($conn, 'DELETE FROM products WHERE id = $1', Array($product_id));
